Prevent recreating review step on every render
The `ReviewContent` component was a simple facade for the
`WPCheckoutOrderReview` component to prevent needing to specify all its
props inline when passed as the `activeStepContent` for the review step.
However, since `ReviewContent` itself was being re-created on every
render, React was unable to recognize that it was the same and re-use
the same actual DOM node; this caused very little visual change since
the resulting DOM was nearly identical, but the input field in the
newly-created DOM node was now different and so the browser lost its
focus.

This change passes the `WPCheckoutOrderReview` element to
`activeStepContent` directly to fix the issue.